high precision duty setting (fixed)

+++ src/Values.cpp
@@ -2,5 +2,5 @@
 
 std::atomic<uint16_t> Values::rpm;
 std::atomic<float> Values::temperature;
-std::atomic<uint8_t> Values::duty;
+std::atomic<float> Values::duty;
 std::atomic<unsigned long> Values::temperatureReadout;
